chore(tests): split tests
